Backend/requirements.txt
Django==3.1.3
numpy==1.26.1
opencv_python==4.7.0.68
Pillow==10.1.0
Pillow==10.1.0
pytesseract==0.3.6
transformers==4.34.1
